Make sure that optimisation is run after scraping
