+++ src/oss4climate/cli.py
@@ -4,12 +4,10 @@ CLI module
 
 from datetime import timedelta
 
-import pandas as pd
 import typer
 
 from oss4climate import scripts
 from oss4climate.scripts import data_publication, listing_search, repository_scraping
-from oss4climate.src.log import log_info, log_warning
 
 app = typer.Typer()
 
@@ -57,6 +55,7 @@ def publish():
 def generate_listing():
     """Generates the updated listing"""
     repository_scraping.scrape_all()
+    repository_scraping.optimise_scraped_data_for_search()
 
 
 @app.command()
@@ -73,39 +72,7 @@ def download_data():
 
 @app.command()
 def optimise():
-    from oss4climate.src.nlp.plaintext import (
-        get_spacy_english_model,
-        reduce_to_informative_lemmas,
-    )
-
-    log_info("Loading spaCy english model")
-    nlp_model = get_spacy_english_model()
-    log_info("- Loaded")
-
-    log_info("Loading input listing")
-    df = pd.read_feather(scripts.FILE_OUTPUT_LISTING_FEATHER)
-    log_info("- Loaded")
-
-    df_opt = df.copy()
-
-    def _f_opt(x: str | None) -> str | None:
-        if x is None:
-            return None
-        try:
-            out = " ".join(reduce_to_informative_lemmas(x, nlp_model=nlp_model))
-        except Exception as e:
-            log_warning(f"Lemmatisation error: {e}")
-            out = "(OPTIMISATION ERROR)"
-        return out
-
-    log_info("Optimising descriptions")
-    df_opt["optimised_description"] = df_opt["description"].apply(_f_opt)
-    log_info("Optimising readmes")
-    df_opt["optimised_readme"] = df_opt["readme"].apply(_f_opt)
-
-    log_info("Exporting input listing")
-    df_opt.to_feather(scripts.FILE_OUTPUT_OPTIMISED_LISTING_FEATHER)
-    log_info("- Exported")
+    repository_scraping.optimise_scraped_data_for_search()
 
 
 if __name__ == "__main__":
